monitoring: exclude Inconsistent resources from drbdDeviceNotUpToDate alert
This alert should report resources in "unexpected" states. However, the way we
handle restoring from a snapshot, or even the addition of new peers to a
resource might be triggering this alert by accident.

The issue is that when new peers are added to an existing volume, it will cause
a partial or full resync, during which the volume stays Inconsistent. This is
expected and fine. We already have a separate alert
"drbdResourceResyncWithoutProgress" that reports disks that are Inconsistent
but made no progress on the resync front.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>