fix "NoSchedule" taints not always working
There was a race condition in the implementation of NoSchedule tolerations
for LinstorSatellites. In particular, it could happen that a Satellite was
removed, but immediately reapplied by LinstorCluster reconciler.

This happens when:
* The LinstorCluster reconciler is running, observing the LinstorSatellite
  still being present.
* The LinstorSatellite is removed.
* The LinstorCluster reconciler reaches the part where it applies the expected
  resources.

Because the LinstorCluster reconciler saw the LinstorSatellite being present,
and because we use server-side apply patches, we did not notice that the
satellite had been removed.

To fix this, ensure that in case we checked for the presence of a satellite,
we also only apply to that specific satellite. Setting the UID of the
to-be-applied resource will case Kubernetes to reject the request if the
satellite with that UID is no longer present.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>