webhook: fix a crash when validating change pool types
When a user updated the storage pool to a different type, we did not properly
protect deep-comparisons of old configuration against the new configuration.
This meant that the webhook would crash when trying to update the storage pool
type, instead of reporting a human-readable error.

Fix that by guarding before making deep storage pool type comparisons and
adding a test to exercise that code path.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>