evacuation: generate user-visible events
Events are temporary resources that are displayed to users when directly
requesting them (kubectl get events), or enhance the output of existing
resources they are related to (for example: kubectl describe pvc).

This makes them a perfect fit to enhance the visibility of the evacuation
process, where PVCs, PVs and Node/Machines are updated or removed by the
Operator.

The new events are generated when:

* A PVC is marked for evacuation.
* A PVC is marked as temporarily schedulable on different nodes.
* A PVC is deleted according to its evacuation-action.
* A PVC is either attached after rescheduling or the attach timeout elapsed.
* A Machine is ready to be drained.
* A Machine is ready to be removed.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>